Refactor home router and update template links; improve HTML structure and CSS inclusion; add base template and extend from it
