+++ main.py
@@ -12,7 +12,7 @@ from users.user_router import router as users_router
 from utils.authentication import router as auth_router
 from matches.match_router import match_router
 # Add WEB Routers
-from web_routers.home_router import index_router
+from web_routers.home_router import router as index_router
 from web_routers.jobposts_router import jobs_router
 from web_routers.user_router import router as user_web_router
