+++ templates/base.html
@@ -1,18 +1,14 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <head>
+    <title>{% block title %}{% endblock %}</title>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="description" content="" />
     <meta name="keywords" content="" />
     <meta name="author" content="Free-Template.co" />
     <link rel="shortcut icon" href="ftco-32x32.png">
-
-    <title>{% block title %}{% endblock %}</title>
-
-    <link rel="stylesheet" href="../static/css/style.css">  
+    
     <link rel="stylesheet" href="../static/css/custom-bs.css"> <!--css/custom-bs.css-->
     <link rel="stylesheet" href="../static/css/jquery.fancybox.min.css"> <!--jquery.fancybox.min.css-->
     <link rel="stylesheet" href="../static/css/bootstrap-select.min.css"> <!--bootstrap-select.min.css-->
@@ -20,42 +16,61 @@
     <link rel="stylesheet" href="../static/fonts/line-icons/style.css"> <!--line-icons/style.css-->
     <link rel="stylesheet" href="../static/css/owl.carousel.min.css"> <!--owl.carousel.min.css-->
     <link rel="stylesheet" href="../static/css/animate.min.css"> <!--animate.min.css-->
+    <link rel="stylesheet" href="../static/css/style.css">
+    
+    {% block extra_css %}{% endblock %}
 
     {% from 'macros.html' import load_header, load_footer %}
 
-</head>
-<body>
-    <body id="top">
 
-        <div id="overlayer"></div>
-        <div class="loader">
-          <div class="spinner-border text-primary" role="status">
-            <span class="sr-only">Loading...</span>
-          </div>
-        </div>
-          
-      
-      <div class="site-wrap">
-      
-          <div class="site-mobile-menu site-navbar-target">
-            <div class="site-mobile-menu-header">
-              <div class="site-mobile-menu-close mt-3">
-                <span class="icon-close2 js-menu-toggle"></span>
-              </div>
-            </div>
-            <div class="site-mobile-menu-body"></div>
-          </div> <!-- .site-mobile-menu -->
-          
-      
-          <!-- NAVBAR -->
-          {{ load_header(user) }}
-            <!-- END NAVBAR -->
-        
-        {% block content %}{% endblock %}
+  </head>
+  <body id="top">
 
-        {{ load_footer() }}
+  <div id="overlayer"></div>
+  <div class="loader">
+    <div class="spinner-border text-primary" role="status">
+      <span class="sr-only">Loading...</span>
+    </div>
+  </div>
 
-        {% block scripts %}{% endblock %}
+
+<div class="site-wrap">
+
+  <div class="site-mobile-menu site-navbar-target">
+    <div class="site-mobile-menu-header">
+      <div class="site-mobile-menu-close mt-3">
+        <span class="icon-close2 js-menu-toggle"></span>
+      </div>
+    </div>
+    <div class="site-mobile-menu-body"></div>
+  </div> <!-- .site-mobile-menu -->
+
+      {{ load_header(get_user(request)) }}
+
+      {% block content %}{% endblock %}
+      
+      {{ load_footer() }}
+
+    </div>
+
+    <!-- JavaScript Files -->
+    <script src="../static/js/jquery.min.js"></script>
+    <script src="../static/js/bootstrap.bundle.min.js"></script>
+    <script src="../static/js/isotope.pkgd.min.js"></script>
+    <script src="../static/js/stickyfill.min.js"></script>
+    <script src="../static/js/jquery.fancybox.min.js"></script>
+    <script src="../static/js/jquery.easing.1.3.js"></script>
     
-</body>
+    <script src="../static/js/jquery.waypoints.min.js"></script>
+    <script src="../static/js/jquery.animateNumber.min.js"></script>
+    <script src="../static/js/owl.carousel.min.js"></script>
+    
+    <script src="../static/js/bootstrap-select.min.js"></script>
+    
+    <script src="../static/js/custom.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap-select@1.13.14/dist/js/bootstrap-select.min.js"></script>
+
+    {% block extra_scripts %}{% endblock %}
+    
+  </body>
 </html>
