+++ templates/home.html
@@ -1,56 +1,8 @@
+{% extends "base.html" %}
 
-<!doctype html>
-<html lang="en">
-  <head>
-    <title>Between Jobs</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="" />
-    <meta name="keywords" content="" />
-    <meta name="author" content="Free-Template.co" />
-    <link rel="shortcut icon" href="ftco-32x32.png">
-    
-    <link rel="stylesheet" href="../static/css/custom-bs.css"> <!--css/custom-bs.css-->
-    <link rel="stylesheet" href="../static/css/jquery.fancybox.min.css"> <!--jquery.fancybox.min.css-->
-    <link rel="stylesheet" href="../static/css/bootstrap-select.min.css"> <!--bootstrap-select.min.css-->
-    <link rel="stylesheet" href="../static/fonts/icomoon/style.css"> <!--icomoon/style.css-->
-    <link rel="stylesheet" href="../static/fonts/line-icons/style.css"> <!--line-icons/style.css-->
-    <link rel="stylesheet" href="../static/css/owl.carousel.min.css"> <!--owl.carousel.min.css-->
-    <link rel="stylesheet" href="../static/css/animate.min.css"> <!--animate.min.css-->
+{% block title %}Home{% endblock %}
 
-    <!-- MAIN CSS -->
-    <link rel="stylesheet" href="../static/css/style.css">  
-
-    {% from 'macros.html' import load_header, load_footer %}
-    
-  
-  </head>
-  <body id="top">
-
-  <div id="overlayer"></div>
-  <div class="loader">
-    <div class="spinner-border text-primary" role="status">
-      <span class="sr-only">Loading...</span>
-    </div>
-  </div>
-
-    
-
-<div class="site-wrap">
-
-    <div class="site-mobile-menu site-navbar-target">
-      <div class="site-mobile-menu-header">
-        <div class="site-mobile-menu-close mt-3">
-          <span class="icon-close2 js-menu-toggle"></span>
-        </div>
-      </div>
-      <div class="site-mobile-menu-body"></div>
-    </div> <!-- .site-mobile-menu -->
-    
-
-    <!-- NAVBAR -->
-    {{ load_header(get_user(request)) }}
-
+{% block content %}
     <!-- HOME -->
     <section class="home-section section-hero overlay bg-image" style="background-image: url('../static/images/hero_1.jpg');" id="home-section">
 
@@ -257,83 +209,5 @@
         </div>
       </div>
     </section>
+    {% endblock %}
     
-    <footer class="site-footer">
-
-      <a href="#top" class="smoothscroll scroll-top">
-        <span class="icon-keyboard_arrow_up"></span>
-      </a>
-
-      <div class="container">
-        <div class="row mb-5">
-          <div class="col-6 col-md-3 mb-4 mb-md-0">
-            <h3>Search Trending</h3>
-            <ul class="list-unstyled">
-              <li><a href="#">Web Design</a></li>
-              <li><a href="#">Graphic Design</a></li>
-              <li><a href="#">Web Developers</a></li>
-              <li><a href="#">Python</a></li>
-              <li><a href="#">HTML5</a></li>
-              <li><a href="#">CSS3</a></li>
-            </ul>
-          </div>
-          <div class="col-6 col-md-3 mb-4 mb-md-0">
-            <h3>Company</h3>
-            <ul class="list-unstyled">
-              <li><a href="#">About Us</a></li>
-              <li><a href="#">Career</a></li>
-              <li><a href="#">Blog</a></li>
-              <li><a href="#">Resources</a></li>
-            </ul>
-          </div>
-          <div class="col-6 col-md-3 mb-4 mb-md-0">
-            <h3>Support</h3>
-            <ul class="list-unstyled">
-              <li><a href="#">Support</a></li>
-              <li><a href="#">Privacy</a></li>
-              <li><a href="#">Terms of Service</a></li>
-            </ul>
-          </div>
-          <div class="col-6 col-md-3 mb-4 mb-md-0">
-            <h3>Contact Us</h3>
-            <div class="footer-social">
-              <a href="#"><span class="icon-facebook"></span></a>
-              <a href="#"><span class="icon-twitter"></span></a>
-              <a href="#"><span class="icon-instagram"></span></a>
-              <a href="#"><span class="icon-linkedin"></span></a>
-            </div>
-          </div>
-        </div>
-
-        <div class="row text-center">
-          <div class="col-12">
-            <p class="copyright"><small>
-              <!-- Link back to Colorlib can't be removed. Template is licensed under CC BY 3.0. -->
-            Copyright &copy;<script>document.write(new Date().getFullYear());</script> All rights reserved | This template is made with <i class="icon-heart text-danger" aria-hidden="true"></i> by <a href="https://colorlib.com" target="_blank" >Colorlib</a>
-            <!-- Link back to Colorlib can't be removed. Template is licensed under CC BY 3.0. --></small></p>
-          </div>
-        </div>
-      </div>
-    </footer>
-  
-  </div>
-
-    <!-- SCRIPTS -->
-    <script src="../static/js/jquery.min.js"></script>
-    <script src="../static/js/bootstrap.bundle.min.js"></script>
-    <script src="../static/js/isotope.pkgd.min.js"></script>
-    <script src="../static/js/stickyfill.min.js"></script>
-    <script src="../static/js/jquery.fancybox.min.js"></script>
-    <script src="../static/js/jquery.easing.1.3.js"></script>
-    
-    <script src="../static/js/jquery.waypoints.min.js"></script>
-    <script src="../static/js/jquery.animateNumber.min.js"></script>
-    <script src="../static/js/owl.carousel.min.js"></script>
-    
-    <script src="../static/js/bootstrap-select.min.js"></script>
-    
-    <script src="../static/js/custom.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap-select@1.13.14/dist/js/bootstrap-select.min.js"></script>
-     
-  </body>
-</html>
