+++ templates/login.html
@@ -35,7 +35,7 @@
                     <div class="site-logo col-6"><a href="home.html">BetweenJobs</a></div>
                     <nav class="mx-auto site-navigation">
                         <ul class="site-menu js-clone-nav d-none d-xl-block ml-0 pl-0">
-                            <li><a href="home.html">Home</a></li>
+                            <li><a href="home">Home</a></li>
                             <li><a href="about.html">About</a></li>
                             <li><a href="contact.html">Contact</a></li>
                         </ul>
