+++ templates/macros.html
@@ -11,14 +11,14 @@
     <header class="site-navbar mt-3">
         <div class="container-fluid">
           <div class="row align-items-center">
-            <div class="site-logo col-6"><a href="home">BetweenJobs</a></div>
+            <div class="site-logo col-6"><a href="/home">BetweenJobs</a></div>
   
             <nav class="mx-auto site-navigation">
               <ul class="site-menu js-clone-nav d-none d-xl-block ml-0 pl-0">
-                <li><a href="home" class="nav-link active">Home</a></li>
+                <li><a href="/home" class="nav-link active">Home</a></li>
                 <li><a href="about.html">About</a></li>
                 <li class="has-children">
-                  <a href="job-listings.html">Job Listings</a>
+                  <a href="jobposts">Job Listings</a>
                   <ul class="dropdown">
                     <li><a href="job-single.html">Job Single</a></li>
                     <li><a href="post-job.html">Post a Job</a></li>
